+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.5] - 2024-12-30
+
+### Added
+
+- Datadog Trace ID support.
+- Hex-encoded Hash support.
+
+### Changed
+
+- Fixed bug: Hashid too long.
+
 ## [0.2.4] - 2024-12-28
 
 ### Added